pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Screenshot OCR application using OpenAI Vision API with global hotkey support"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "keyboard>=0.13.5",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "pygetwindow>=0.0.9",
    "pyperclip>=1.8.2",
    "streamlit>=1.52.2",
    "uvicorn>=0.24.0",
]

# Note: PaddleOCR is NOT included in dependencies to avoid conflicts with paddlepaddle-gpu.
# If paddleocr is listed here, package managers will install paddlepaddle (CPU version),
# which will replace paddlepaddle-gpu if it's already installed.
#
# For GPU support with PaddleOCR, install paddlepaddle-gpu==2.6.2 separately FIRST:
# 
# Check your CUDA version first: nvidia-smi
# 
# PyPI only provides CUDA 10.2 builds. For other CUDA versions, use the official website:
# Reference: https://pypi.org/project/paddlepaddle-gpu/
# 
# For CUDA 10.2 (from PyPI):
#   pip install paddlepaddle-gpu==2.6.2
# 
# For CUDA 11.2/11.6/11.7 (from PaddlePaddle website):
#   Check the official installation document: https://www.paddlepaddle.org.cn/install/quick
#   Example for CUDA 11.8 (if available):
#   pip install paddlepaddle-gpu==2.6.2 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/
# 
# For CPU-only:
#   pip install paddlepaddle==2.6.2
# 
# After installing paddlepaddle-gpu, install PaddleOCR:
#   pip install paddleocr
#
# IMPORTANT: Install paddlepaddle-gpu BEFORE paddleocr to ensure the GPU version
# is used. If you install paddleocr first, it will pull in paddlepaddle (CPU) as
# a dependency, which will conflict with paddlepaddle-gpu.
#
# The application gracefully handles missing PaddleOCR - it will process the full
# image instead of detecting text regions if PaddleOCR is not installed.
#
# See README.md for detailed installation instructions
